backend/pyproject.toml
[project]
name = "zenboard"
version = "0.1.0"
description = "Zenith - 极简团队协作系统后端"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "cryptography>=45.0.6",
    "fastapi>=0.116.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.35.0",
    "alembic>=1.16.1",
    "httpx>=0.28.1",
    "pytz>=2025.2",
    "passlib>=1.7.4",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
]

[tool.uv]
python-downloads = "never"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true
